requirements.txt
Flask==3.0.0
gunicorn==21.2.0
ultralytics>=8.3.0
Pillow==10.0.0
opencv-python-headless>=4.8.0
numpy>=1.26.0
werkzeug==3.0.0